# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "VICAM API",
    "version": "0.3.0",
    "description": "Operational contracts for the VICAM modular monolith."
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "ErrorEnvelope": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "minLength": 1
          },
          "message": {
            "type": "string",
            "minLength": 1
          },
          "fieldErrors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          "requestId": {
            "type": "string",
            "minLength": 1,
            "example": "01JABCDEF0123456789ABCDE"
          },
          "details": {
            "type": "object",
            "additionalProperties": {}
          }
        },
        "required": ["code", "message", "requestId"]
      },
...
      "PaginationMeta": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer",
            "exclusiveMinimum": 0
          },
          "pageSize": {
            "type": "integer",
            "exclusiveMinimum": 0,
            "maximum": 100
          },
          "total": {
            "type": "integer",
            "minimum": 0
          },
          "totalPages": {
            "type": "integer",
            "minimum": 0
          }
        },
        "required": ["page", "pageSize", "total", "totalPages"]
      },
...
      "NotificationsPage": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Notification"
            }
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationMeta"
          }
        },
        "required": ["items", "pagination"]
      },
      "Notification": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "type": {
            "type": "string",
            "minLength": 1,
            "maxLength": 80
          },
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "body": {
            "type": "string",
            "minLength": 1,
            "maxLength": 500
          },
          "resourceType": {
            "type": ["string", "null"],
            "maxLength": 80
          },
          "resourceId": {
            "type": ["string", "null"],
            "format": "uuid"
          },
          "readAt": {
            "type": ["string", "null"],
            "format": "date-time"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "type",
          "title",
          "body",
          "resourceType",
          "resourceId",
          "readAt",
          "createdAt"
        ]
      },
      "PushSubscriptionRequest": {
        "type": "object",
        "properties": {
          "deviceId": {
            "type": "string",
            "format": "uuid"
          },
          "endpoint": {
            "type": "string",
            "maxLength": 2000,
            "format": "uri"
          },
          "p256dh": {
            "type": "string",
            "minLength": 1,
            "maxLength": 1000
          },
          "auth": {
            "type": "string",
            "minLength": 1,
            "maxLength": 1000
          }
        },
        "required": ["deviceId", "endpoint", "p256dh", "auth"]
      },
...
      "AppSettings": {
        "type": "object",
        "properties": {
          "offlineWindowHours": {
            "type": "integer",
            "minimum": 1,
            "maximum": 72
          },
          "visitReminderOffsetsMinutes": {
            "type": "array",
            "items": {
              "type": "integer",
              "exclusiveMinimum": 0
            },
            "maxItems": 5
          },
          "taskReminderOffsetsMinutes": {
            "type": "array",
            "items": {
              "type": "integer",
              "exclusiveMinimum": 0
            },
            "maxItems": 5
          },
          "supervisorReportsEnabled": {
            "type": "boolean"
          },
          "documentLimitBytes": {
            "type": "integer",
            "exclusiveMinimum": 0,
            "maximum": 10485760
          },
          "defaultTimezone": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "retentionDays": {
            "type": "object",
            "properties": {
              "exports": {
                "type": "number",
                "enum": [7]
              },
              "documentsTrash": {
                "type": "number",
                "enum": [30]
              },
              "jobs": {
                "type": "number",
                "enum": [90]
              }
            },
            "required": ["exports", "documentsTrash", "jobs"]
          },
          "version": {
            "type": "integer",
            "exclusiveMinimum": 0
          }
        },
        "required": [
          "offlineWindowHours",
          "visitReminderOffsetsMinutes",
          "taskReminderOffsetsMinutes",
          "supervisorReportsEnabled",
          "documentLimitBytes",
          "defaultTimezone",
          "retentionDays",
          "version"
        ]
      },
      "UpdateAppSettingsRequest": {
        "type": "object",
        "properties": {
          "offlineWindowHours": {
            "type": "integer",
            "minimum": 1,
            "maximum": 72
          },
          "visitReminderOffsetsMinutes": {
            "type": "array",
            "items": {
              "type": "integer",
              "exclusiveMinimum": 0
            },
            "maxItems": 5
          },
          "taskReminderOffsetsMinutes": {
            "type": "array",
            "items": {
              "type": "integer",
              "exclusiveMinimum": 0
            },
            "maxItems": 5
          },
          "supervisorReportsEnabled": {
            "type": "boolean"
          },
          "documentLimitBytes": {
            "type": "integer",
            "exclusiveMinimum": 0,
            "maximum": 10485760
          },
          "defaultTimezone": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "version": {
            "type": "integer",
            "exclusiveMinimum": 0
          }
        },
        "required": [
          "offlineWindowHours",
          "visitReminderOffsetsMinutes",
          "taskReminderOffsetsMinutes",
          "supervisorReportsEnabled",
          "documentLimitBytes",
          "defaultTimezone",
          "version"
        ]
      }
    },
...
  },
  "paths": {
...
    "/push-subscriptions/{id}": {
      "delete": {
        "tags": ["Notifications"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "id",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 200
            },
            "required": false,
            "name": "idempotency-key",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "Current device push subscription removed.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    }
                  },
                  "required": ["id"]
                }
              }
            }
          },
          "401": {
            "description": "Authentication required or session expired.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient role or ownership.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Resource does not exist or is not visible to the caller.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Version or state conflict.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Business validation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "Service unavailable or feature disabled by an operational rollback flag.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
    "/notifications": {
      "get": {
        "tags": ["Notifications"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            },
            "required": false,
            "name": "page",
            "in": "query"
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 15,
              "maximum": 15,
              "default": 15
            },
            "required": false,
            "name": "pageSize",
            "in": "query"
          },
          {
            "schema": {
              "type": ["boolean", "null"]
            },
            "required": false,
            "name": "unread",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Notifications in the current user's scope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationsPage"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required or session expired.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/notifications/read-all": {
      "post": {
        "tags": ["Notifications"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 200
            },
            "required": false,
            "name": "idempotency-key",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "Notifications marked as read.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "updated": {
                      "type": "integer",
                      "minimum": 0
                    }
                  },
                  "required": ["updated"]
                }
              }
            }
          },
          "401": {
            "description": "Authentication required or session expired.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/notifications/{id}/read": {
      "post": {
        "tags": ["Notifications"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "id",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 200
            },
            "required": false,
            "name": "idempotency-key",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "Notifications marked as read.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Notification"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required or session expired.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/push-subscriptions": {
      "post": {
        "tags": ["Notifications"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 200
            },
            "required": false,
            "name": "idempotency-key",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PushSubscriptionRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Push subscription stored.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    }
                  },
                  "required": ["id"]
                }
              }
            }
          },
          "422": {
            "description": "Business validation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
    "/settings": {
      "get": {
        "tags": ["Settings"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Typed application settings.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppSettings"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient role or ownership.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": ["Settings"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 200
            },
            "required": false,
            "name": "idempotency-key",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAppSettingsRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Settings updated and audited.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppSettings"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient role or ownership.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Version or state conflict.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    }
  },
...
}